Fix GitLab CI + small improvements (#202)
This MR fixes the following issues encountered by the internal CI:
- the LPF tests could fail on gitlab runners that are not large enough;
- some CI coverage tests were skipped due to the missing CMake `DATASET_DIR`
- compilation of the SparseBLAS was always attempted no matter which backends were enabled
- fix sporadic bug of Coverage with OpenMP.

What is a "large enough" runner is undefined-- the documentation instead now mentions the configuration used for our internal CI, where this CI pipeline consistently passes, and thus describes what is sufficiently large.

The selection of a large-enough runner *may* require a specific (set of) runner tags that depend on your CI configuration. This MR allows for such tags to be made known via a site-specific (fixed global) variable, `LPF_PREFERRED_RUNNERS`, which may be set within your GitLab CI configuration. This way, new GitLab CI deployments should work out-of-the-box if the runners already happen to be "large enough", while if the explicit selection of "large-enough" runners is required, it is now easy to do so.

Thanks to @byjtew and @alberto-scolari for contributing this MR!