650 Potential bug in bps1d (#232)
Benjamin found a bug where a BSP1D matrix would not initialise the ID of the sub-backends properly-- it would always assign an invalid ID. There is no known user-facing issue because of this bug-- when retrieving a container ID, the "front" backend is queried for it, not the sub-backend.

Nonetheless, without the fix this MR includes, it might have caused issues in some rare or future conditions. The MR also includes a new unit test that checks for all distributed backends (presently, `bsp1d` and `hybrid`), that the sub-backend IDs are correctly and consistently instantiated. It also fixes a meta-bug in one of the assertions checked for during `BSP1D` matrix construction.

Many thanks Benjamin for finding and fixing this bug!