260 bicgstab smoke tests sometimes does not report useful time 1 (#262)
This MR addresses the following:

as with issue #166, but now also detected for the BiCGstab smoke test instead of the CG one, if the BiCGstab solver fails to converge (and thus returns the `grb::FAILED` error code), the useful time is not recorded by the smoke test.

This MR fixes the issue for BiCGstab, and reviews the smoke tests for all iterative solvers so to make sure the same issue does not persist elsewhere. This MR additionally makes uniform the implementation of all smoke tests for iterative solvers, and makes the maximum number of iterations a command-line argument for all iterative solver smoke tests.

As always, it also includes various minor code style fixes.