264 repeated performance tests should skip repeated file io (#265)
This MR resolves the first item of issue #264, and addresses the ALP performance testing setup, in particular all performance tests that operate on input files.

The situation before this MR is that each experiment also tests I/O performance. For large input files, this drives up the required performance test time, as I/O tends to be by far the slowest component in every benchmark. Yet when running e.g. a performance test for BiCGstab, one is typically interested in the performance of BiCGstab, not in the performance of I/O. With "experiments", we here mean here the so-called "outer iterations" of the many performance tests included with ALP.

This MR modifies the performance testing framework to perform file I/O only once. During this initial I/O phase, the file contents are buffered in memory. The actual experiments will then repeatedly read from the in-memory file storage. We do not simply load and parse the file to an ALP container once and then keep reusing the same parsed containers, as on some platforms, the precise allocation (which can't always be precisely controlled) may have significant effect on performance-- and this is exactly the kind of effect that having a single benchmark execute the same experiment multiple times was designed to measure.

This MR also, as a bonus, includes
1. code style fixes in the Pregel interface: `num_vertices` -> `numVertices`, `num_edges` -> `numEdges`, `get_matrix` -> `getMatrix`. **This change does affect ALP/Pregel user code!**
2. a singleton utility class has been added,
3. code style review of all smoke and performance tests, as well as,
4. the usual smaller code style fixes in unrelated files. In particular, some filenames erroneously employed CamelCase naming, while they should have used camelCase (or maybe, rather, dromeDary).