275 matrix file iterator should include properties by reference (#277)
The MatrixFileIterator for void-valued nonzero types copied-by-value the matrix properties. It should instead hold a reference to those properties. This manifested as a performance bug in `develop` after MR #264 that caused our internal CI to take more than 2 hours on a cursory performance test suite, instead of roughly 40 minutes.

MR #264 itself contained no issues-- all it did was include a programming pattern that kept re-creating an end-iterator, which, in turn, kept creating deep copies of the maps in the matrix file property class. Those maps furthermore are only populated when in indirect parsing mode.

This MR fixes the bug and includes code style fixes to the matrix parser files.