640 Provide Vector( std::initializer_list<T> ) constructor (#227)
This MR proposes a constructor from an STL initializer-list for `grb::Vector`s. It provides an implementation for all backends, and includes a unit test for the new functionality. This new constructor is specified to operate using sequential I/O; indeed, in its intended use, the initializer list itself will be replicated across user processes. The constructed `grb::Vector` will have a size equal to the number of elements in the initializer list (that is, the result is a dense vector).

The MR includes some unrelated code style fixes, and some other unrelated minor fixes to suppress warnings that appeared with GCC 9.3.1.

Many thanks to Benjamin for providing this new feature!