259 nonblocking backend uses a dynamic schedule (#292)
Before this MR, some OpenMP schedules elected by the nonblocking backend were `dynamic` even though the workload it parallelises was uniform. The overarching ALP principle is to, for uniform workloads and in conjunction with OpenMP, to elect a static schedule instead. This MR applies this fix.

The MR also includes an unrelated minor documentation fix, and some minor code style fixes.

Thanks to both Alberto and Aristeidis for both having detected this issue!