Issue #293: fixing generation of CMake configuration files (#299)
When issuing make install, CMake creates several CMake configuration files for CMake-based projects to easily import ALP, as from the docs. Some include paths (for transition libraries), however, and prior to this MR, are wrong in the installation, due to wrong settings in include/CMakeLists.txt. This MR corrects this issue.

Many thanks to Alberto for detecting and fixing this issue!