Wrong values printed in debug output (#298)
In a testing utility that checks for equality between two matrix structures, if inequality was found, the wrong data was printed to screen. This meta-bug is not expected to have had any ill effects -- the error detection mechanism itself was correct (and for everything that has been merged or released did not trigger); nonetheless, this MR fixes it.

Many thanks to Aleksa for detecting and fixing the bug!