Issue #261: adding missing dense descriptor and delete unnecessary se… (#288)
Prior to this MR, the PageRank implementation does not pass the dense descriptor when it could, while it includes the use of the expensive `invert_mask` in its performance-critical loop. Mask inversion is expensive since it requires iterating over the full vector range, while, in this case, the inverted mask is constant. Additionally, an unnecessary call to `grb::set` preceded a call to the out-of-place `grb::eWiseApply`. Hence, this MR introduces the following changes:
- the inverted mask is precomputed once, while in the performance-critical loop the precomputed inversion is reused as a structural mask;
- the PR algorithm is reviewed to pass the dense descriptor whenever possible;
- the PR algorithm no longer makes the detected unnecessary call to `grb::set`.

This improvement does come at the cost of one additional work-space vector that is now required for the PageRank implementation.

In implementing these fixes, the following feature has been added:
- the `grb::eWiseLambda` now takes a descriptor template argument. This was previously accepted in the nonblocking backend, but was missing from the specification as well as missing in the implementations of other backends.

The following bugs were encountered and fixed with this MR:
- mixed-domain vector-to-scalar reductions were bugged for the nonblocking backend: when the second (underlying) monoid domain type did not match its third (output) domain type, a compiler error would be returned. This MR fixes the behaviour for both left- and right-handed reductions;
- the descriptors for multiple-vector eWiseLambdas were not correctly propagated for the nonblocking backend.

Thanks to Aristeidis and Benjamin for detecting and fixing a variety of issues this MR addresses!