633 Provide grb::foldl/foldr( [in,out] T&, [in] Matrix<D>, monoid ) (#188)
This MR specifies and implements matrix-to-scalar reduction routines. The matrix may be input masked (by a separate, second given matrix). The reductions take either a commutative monoid or a semiring -- if the latter, its additive monoid will be used for reduction (which, by definition of a semiring, is also commutative). The monoid variants cannot handle pattern matrices, and cannot handle the `add_identity` descriptor-- for these, the "one" identity of a semiring is necessary.
Since with ALP/GraphBLAS, operators, from which monoids and semirings are comprised, may have different input and different output domains (i.e., `A x B -> C`), this MR introduces left-looking and right-looking variants of this reduction operation and calls them `foldl` and `foldr`, respectively. This is in line with the previously already included primitives for scalar-to-scalar and vector-to-scalar reduction.

The MR introduces exhaustive unit testing for the new functionality, and, as always, includes code style fixes in files related and unrelated to this new feature. Notable other inclusions:
- an internal utility has been introduced that helps asserting return codes are synchronised across multiple user processes;
- a systematic way to derive index-translation lambdas for so-called "union coordinates" has been introduced. These lambdas are by convention used as a mechanism to let the BSP1D backend (or any distributed backend) translate process-local calls to any single-process final backend. This new feature may yet be further improved-- see GitHub issue #322 .

Many thanks to Benjamin for the implementation and testing for this one!