a selection operator (#271)
This MR introduces the matrix selection feature -- given a operator that takes matrix nonzero coordinates and a value as arguments and produces a boolean, elements from an input matrix are selected to be copied in some output matrix. This primitive operates in an out-of-place fashion. In-place semantics may be achieved via a temporary (pattern) matrix and fold, and thus are currently not planned to be added. Standard matrix selection operators such as (strictly) lower triangular etc. are provided, including type traits for them. Non-standard selection operators provided by users may be employed also since the new grb::select primitive accepts arbitrary lambda functions as a selection operator as well.

This MR also fixes a bug for the BSP1D and hybrid eWiseApply on matrices with the resize mode, which returned `SUCCESS` even if one or more processes encountered an issue (such as out-of-memory conditions). The MR includes unit tests for matrix distributions and the new select operator, as well as code style fixes in related and unrelated files.