Fix a performance bug in the nonblocking backend (#326)
Since #292 , a performance bug has snuck in for the nonblocking backend that causes it to underperform for small-to-mid sizes matrices (up to size 1.5 million). This bug was not present in v0.7. It had been present in the `develop` branch from MR #292 up to this MR (#326).

For larger matrices, the bug stifled performance of the nonblocking backend, causing it to be on par with or exceed the`reference_omp` backend, but not by as much as it should have.

This MR fixes the bug, which was found to be the following: since a tile consists of many elements, the introduced static schedule should use a blocksize of one, and not the default larger blocksize. Taking a larger blocksize in fact simulates (in a suboptimal way) the use of nonblocking execution with a cache size that is multiplied with the blocksize. With this MR, the blocksize is explicitly set to one instead.