655 Provide grb::set( matrix, mask, value ) (#246)
This MR formalises the `grb::set(matrix,matrix)` and `grb::set(matrix,mask,scalar)` variants. For the former, all backends did implement it and an exhaustive unit test existed prior to this MR, but a specification of the function was missing-- this MR adds this. For the latter, all backends did implement one variant of it which was tested in the unit test for the matrix-matrix `grb::set`. This MR specifies and implements all possible variants of `grb::set(matrix,mask,scalar)`, and adds a unit test that exhaustively tests all combinations under which `grb::set(matrix,mask,scalar)` may be called.

To realise this MR, the following features were introduced:

- the ALP utilities now include a prefix-sum algorithm, both sequential and in OpenMP. The implementation of some of the new functionalities depend on this operation. Existing prefix-sum computations should rely on this new utility function; GitHub issue #332 is to follow up on that.
- ALP objects now include ALP types such as `grb::Phase`. This was required to unambiguously dispatch between `grb::set(matrix,matrix)` and `grb::set(matrix,mask,scalar)` via SFINAE.

Additionally, this MR includes the following bugfixes:

1. copy-contructor of pattern matrix iterators for more than one user processes did not copy PID information correctly;
2. the `fold_matrix_to_scalar` unit test relied on undefined behaviour, here fixed;
3. matrix reduction into a scalar could fail if the matrix is small w.r.t. the number of user processes;
4. selection operators erroneously did not appear in user documentation;
5. the specification for resize and execute phases was outdated: if in execution phase it is detected that the output container has insufficient capacity, then primitives should return `ILLEGAL`;
6. the BSP1D set-vector-to-value erroneously returned `FAILED` instead, herewith fixed.

The impact of bugs 2-4 is expected very low as it was only present in the `develop` branch and only for a limited time. Issues 1 and 5 were present longer, however, prior to crafting this MR and as far as we know, seem never to have been triggered or noticed. Issue 5 furthermore did not induce any code changes, except for the bug in issue no. 6. That bug (no. 6), however, never triggers in the currently available back-ends because vectors have fixed full capacities (the bug would only apply for final backends that do not have such a fixed full capacity).

Last, and as always, the MR minor includes code style and documentation fixes in files related and unrelated to this MR.

Many thanks to Benjamin Lozes @byjtew for drafting the initial version of the MR!