249 alp passes 0 as max calls to lpf collectives initialisation (#302)
This MR fixes a bug where `lpf_coll_t` was not initialised properly. This bug never resulted in a noticeable issue, since the present implementation of the LPF collectives ignores the erroneous argument that ALP was passing in. The nevertheless not-up-to-spec initialisation of LPF collectives happened repeatedly throughout the code base. This MR totally overhauls the implementation of the collectives to address the issue.

Within LPF, there are two classes of collectives: the one exposed via the public API `grb::collectives<>`, and raw internal collectives that are used in the implementation of e.g. the `BSP1D` and `hybrid` backends. Prior to this MR, the implementations of these collectives were disjoint, while with this MR, both types of collectives are managed by a shared context (internally encapsulated by the `BSP1D_Data` object). This central management means every process maintains but a single `lpf_coll_t` instance that is reused every time a public or internal collective is called.

The revision of the collectives also now includes a systematic choice where collectives on scalars are orchestrated through the global buffer so as to prevent having to register the scalar memory addresses every time a collective is called. This is likely to cut latency costs significantly compared to the status prior to this MR. For collectives on arrays, the source and/or destination arrays will be registered which incurs a latency costs-- but a cost that, assuming large enough array sizes, is much lower than that of having to copy the payload to and from the global buffer. 

Features this MR introduces in realising the refactored collectives:
- a notion of a final backend, which is the final backend a composed backend relies on. The FinalBackend class collects a set of minimal and low-level functions that all final backends should provide, such as memcpy; and
- a utility that translates LPF error codes into corresponding ALP error codes.

Other issues this MR addresses:
- BSP1D did not seem to handle some border cases of `grb::setElement` correctly -- these would trigger only in out-of-memory conditions, or in erroneous use of the primitive, so impact of the bug is expected to be minor;
- the `grb::setElement` and `grb::select` base specifications did not list all error conditions or did not do so correctly, herewith fixed;
- a metabug in `grb::select` had in `_DEBUG` mode an attempt to print an iterator to `stdout`, instead of first dereferencing it-- herewith fixed.

As always, the MR includes both related and unrelated code style fixes.