Issue #303: framework for reference_dense added
This commit comes from a split of the original commit whose details are as follows:
SHA1 ID: a7c1db0eb45326184dadec39a6a7731eb16a39fd
Author: Albert-Jan Yzelman <yzelman@linux.fritz.box>  2022-01-14 15:52:26
Committer: Albert-Jan Yzelman <yzelman@linux.fritz.box>  2022-01-30 13:48:35

The reason for splitting is to separate the changes irrelevant to the branch 303 into a separate branch.
The split produced 3 separate commits.
This commit is number 1.
This commit in particular is relevant to 303 and will remain in it.