333 current parallel spgemm cannot be merged (#334)
Concept release notes:

This MR provides a novel implementation for the sparse matrix--sparse matrix multiplication, `grb::mxm`. It modifies the specifications in the following ways:

 1. the `mxm` is now in-place, meaning, it computes C += AB. This was intended a long time ago, but somehow the `mxm` was never adapted until now. If the old behaviour is desired, first clear C (`grb::clear`) before calling `mxm`.

 2. the `grb::resize` now retains old nonzeroes, and (therefore) calls to resize must request a capacity that is larger (or equal) to the current number of values in the container, or otherwise `grb::ILLEGAL` shall be returned.

The newly implemented mxm uses as many threads as internal buffers allow. It still follows Gustavson's approach, where now every thread employs its own private sparse accumulator (SPA)-- which is why the size of the global buffer determines the parallelism of the approach. At least one SPA is guaranteed as the first thread will use the output matrix buffers for its SPA-- threads beyond that first one will use any remaining space in the global buffer. Note that the global buffer size is Theta(m+n+nz) for any matrix in the ALP context, while the SPA memory requirements are proportional to n (or m, if transposed)-- therefore, if nz ~ cn, then this new implementation allows the `mxm` to proceed with Theta(c) threads.

Regarding the `resize`, recall that the old behaviour was to clear the given container. The rationale for change this behaviour is that it became apparent these semantics make it both awkward to use as well as implement the `RESIZE` phase of the `mxm`.

This awkwardness was never observed for level-{1,2} primitives because all current implementations use a fixed full capacity for vector data.

Some related specification clarifications were added to that of the `resize`: whether backends allow for shrinking memory usage after a call to `resize` is defined by its performance semantics. The functional semantics leave this behaviour as optional (and current implementations do not ever shrink capacities).

This MR also includes the following bugfixes:
 - [functional bug] the `RESIZE` phase for the masked `grb::set` for matrices sometimes did not resize to a large-enough capacity
 - [functional bug] in the binsearch utility. Expected impact is nil as the function presently is not used from within ALP
 - [performance bug] the dense descriptor in `grb::set(vector,vector,vector)` could erroneously be ignored
 - [performance bug] the analytic model in the `grb::set` for matrices could erroneously select way too many threads, lowering performance significantly
 - [performance bug] same as the directly preceding item, but for the OpenMP prefix sum utility
 - [performance bug] same as the preceding items, but for `grb::select`
 - [performance bug] the fold-matrix-to-scalar unit test used copy-by-value instead of const-reference capture. This has no effect on user code.

This MR also includes the following new features:
 - [utils] a utility that determines the maximum argument for which a monotonic decreasing function on a given range returns its maximum is provided
 - [utils] an OpenMP prefix-sum variant that also shifts all entries one position to the right has been provided
 - [utils] an unordered-memmove has been implemented, with both sequential and OpenMP parallel variants (used by the new in-place `mxm`)
 - [debugging] more targeted debug flags (e.g., `_DEBUG_UTILS_PREFIXSUM` to debug only that particular utility or `_DEBUG_REFERENCE_BLAS3` to debug only level-3 primitives in the reference backends. Defining `_DEBUG` will still enable all debug outputs)
 - [native] a non_owning_view descriptor has been introduced as a requirement for an ALP native interface. This descriptor indicates that containers are used whose associated memory is not managed by ALP.
 - [internal] The coordinates class now supports calling a parallel set or a sequential one (for use as thread-local SPAs)

Tests have been modified as follows:
 - [unit] the SpMSpM multiplication unit test now also tests in-place behaviour
 - [unit] the SpMSpM multiplication unit test now also tests the force_row_major descriptor, which now appears as a separate test in the test suite
 - [unit] the capacity unit test has been brought in-line with the updated specifications of `grb::resize`
 - [unit] the id test now uses 100 tries instead of a 1000 for CI performance reasons
 - [unit] the parallelRegularIterators test implicitly and unintendedly cast a double to int
 - [performance] the SpMSpM multiplication driver can now benchmark CRS only also

As always, this MR also includes some code style fixes and provides some missing code documentations.