Add early exit in preconditioned CG when initial residual is below tolerance (#350)
Previously, the (preconditioned) Conjugate Gradient (CG) algorithm applies the preconditioner before checking whether the current residual already met the required tolerance. The changes from this MR allows the CG to return immediately when the initial residual is small enough, avoiding the unnecessary cost of a preconditioner call. This change can improve performance in cases where the initial guess is already close to the solution. Thanks to @ChristosMatzoros for proposing this enhancement!