360 transition path allocates workspace vector data for sparse information (#361)
Previous to this MR the solver transition path allocates workspace vector data using the standard ALP vector constructors, which results allocating a sparse accumulator (SPA). SPAs are used for ALP/GraphBLAS vectors to handle sparsity efficiently. However, in the use case of exactly this solver transition path, the workspace vector data is guaranteed dense, which results in unused memory. This clashes with a use case where as many independent solves as possible are to be handled simultaneously, which actually results in a memory-constrained scenario. To effectively support this use case, this MR uses the transition path vector wrappers to, instead of using the standard ALP/GraphBLAS vector constructors, wrap around a given memory region instead. This memory region is of size n times `sizeof(T)`, where n is the vector size and `T` is the vector element type, thus allocating precisely what is required for a dense vector and not more.

Additionally, the solver transition path API now can take hints when creating a PCG solver handle that control how much and how the memory should be allocated: with or without NUMA support (in the aforementioned memory constrained many-solver scenario, a local allocation policy may be preferred), and with or without preconditioner support (preconditioned solves require one extra workspace vector to be allocated). These are hints: the specification of the transition path allows implementations to ignore hints. The present ALP implementation introduced with this MR only ignores a requested NUMA-aware allocation if it was configured without `libnuma` support-- in all other cases the hints are respected. The "hint-semantic" also allows this MR's implementation to still execute a preconditioned solve successfully, even if a hint was given that this would not happen. In this situation, on-demand allocation using ALP/GraphBLAS' standard vector allocation mechanism will still be used, in which case unused memory will still be allocated. Additionally, in this case, a warning is printed to stderr.

While options are now possible, the standard constructors available prior to this MR are still present. These will default to NUMA-aware allocations with preconditioning, which is the case closest to the constructor as it was and also reflective of the most common use case.

Finally, this MR allows for users to supply their own workspace as a pointer-size pair. This supports perhaps even more exotic use cases where the user program has their own non-standard allocation schemes -- or, if applications would want to reuse buffers between different phases of their application, again when in a memory-constrained setting.

Thanks to @ChristosMatzoros for integration with this particular use case, testing the fixes, as well as the final code review!