Fixes static-assert failure from unused code path #371 (#372)
@djelovina observed it was possible for a static assertion from the sparse code path of `grb::eWiseApply` to trigger while passing in the `dense` descriptor to that primitive. This MR ensures that the corresponding sparse code path's static assert now does *not* trigger when the `dense` descriptor is given, using the "C++11 way" of ensuring so. Issue #373 has been raised to handle it more elegantly using the C++17 `if constexpr`.

Additionally, in handling this MR it was found that a similar static-assert was missing from the dense code path for `grb::eWiseApply`, and has been added. The MR also enhances the warnings to the `right_assign_if` and `left_assign_if` operators regarding their use in out-of-place primitives which is what led to detecting this bug.

Thanks to Denis for finding and reporting the issue, as well as for help reviewing this fix.