168 Added copy-assignment vector ID unit tests (#382)
Prior to this MR, the implementation guarantees retention of IDs when subject to copy-assignment. However, under move-assignment, the moved-to container relinquishes its resources, taking over the resources of the moved-from container instead. This implies that the moved-to container takes over the ID of the moved-from container (while the ID, and general state, of the moved-from container becomes undefined). This issue was previously detected and fixed for ALP/GraphBLAS matrices in #279 / #286, which also introduces a unit test for preventing regressions.

This MR introduces a unit test for move-assignment of vectors, which confirmed that ALP prior to this MR had invalid behaviour re IDs when move-assigning vectors, as issue #168 indeed suspected. This MR fixes those issues, and includes two minor additional code style fixes.