345 Compilation warnings suppression (#379)
This MR removes all compiler warnings under GCC 9.4.0, 11.4.0, and 11.5.0. Thanks to Giovanni Gaio for contributing this fix, as well as the preceding MR!

This MR additionally includes a hotfix to a metabug: one of the performance tests used the dense descriptor while not all input vectors were dense, herewith fixed. This bug was only detected by the nonblocking backend, which would furthermore only detect it when `grb::wait()` was called while 1) a pipeline was still awaiting execution and 2) previously inputs were not all materialised yet-- i.e., it would only be detected as part of a type of race condition -- "type of" since no erroneous outputs could materialise.

As often, this MR also includes some unrelated code style fixes.