255 refactor `alloc.hpp` (#383)
The `alloc.hpp` started out as a utility, but currently the code base requires that each backend provides its own allocator functions. This has two reasons:
 - some backends are shared-memory parallel and require NUMA-aware allocation routines;
 - some backends target embedded or accelerator architectures and require specific allocations also.

Therefore, `alloc.hpp` should be moved from the `utils` directory to the base directory, and follow the same pattern as e.g. the collectives to enable back-end specific implementations. This MR implements the required changes, and also revises the documentation of the allocation functions.

Thanks to Giovanni Gaio for providing the fix!