151 benchmarker base does not check for errors in collective calls (#384)
MR to fix an issue that was already largely fixed in the previous #235 / #239 . This MR adds more graceful exits on failures (as opposed to calling abort). It also fixes a logic error that could result in printing the wrong error message (sleep interrupted) while actually a different error was encountered. Finally, it fixes a missing include.

Thanks to Giovanni Gaio for looking into the issue and providing the MR!